requirements.txt
streamlit
pandas
numpy
scikit-learn
xgboost
imblearn
plotly
joblib
kafka-python
streamlit-autorefresh